Improve cron action for session cleanup

+++ modules/Vtiger/crons/SessionCleaner.php
@@ -8,6 +8,7 @@
  * @license   YetiForce Public License 3.0 (licenses/LicenseEN.txt or yetiforce.com)
  * @author    Tomasz Kur <t.kur@yetiforce.com>
  * @author    Mariusz Krzaczkowski <m.krzaczkowski@yetiforce.com>
+ * @author    Radosław Skrzypczak <r.skrzypczak@yetiforce.com>
  */
 
 /**
@@ -22,6 +23,9 @@ class Vtiger_SessionCleaner_Cron extends \App\CronHandler
 	{
 		$dbCommand = \App\Db::getInstance('webservice')->createCommand();
 		foreach (\Api\Core\Containers::$listTables as $row) {
+			if (!isset($row['session'])) {
+				continue;
+			}
 			$dbCommand->delete($row['session'], ['<', 'created', date('Y-m-d H:i:s', strtotime('now') - \Config\Security::$apiLifetimeSessionCreate * 60)])->execute();
 			$dbCommand->delete($row['session'], ['<', 'changed', date('Y-m-d H:i:s', strtotime('now') - \Config\Security::$apiLifetimeSessionUpdate * 60)])->execute();
 		}
